Moeller Trumbore, also add a t_min to the hit routine

+++ BVHBuilder.h
@@ -9,6 +9,6 @@ class Hitable;
 class BVHBuilder
 {
 public:
-	static std::shared_ptr<BVHNode> buildBVH(std::vector<std::shared_ptr<Hitable>>& list, int offset, int size, float time0, float time1);
+	static std::shared_ptr<BVHNode> buildBVH(std::vector<std::shared_ptr<Hitable>>& list, size_t offset, size_t size, float time0, float time1);
 };
